[#121] harden the eslint config
This commit adds:
* the react-hooks plugin to the configuration
* the eqeqeq to the configuration

Also two issues regarding the react-hooks were treated:

Normally it is impossible to call a hook in a conditional branch (if,
loops, ...). That is because it would mess with how React stores it's
data. https://react.dev/reference/rules/rules-of-hooks However in this
specific configuration, the number determining how many loops are
going to occur gets defined by the number of pods within the deployment.
This means this number is guaranteed to stay the same while the page is
displayed, as long as there is no support for the hot reloading of the
deployment details.

/!\ I ran the linter with `yarn lint` and some files got modified along
the way.