jolokia: using useEffect for making queries
This commit changes the way the network requests are performed.

Before, requests we made directly in the JSX Components, thus exposing
the application to side effects during render time. This was changed to
use useEffect to encapsulate the network calls. As the useEffect is
scheduled after the rendering time, the time it takes to actually
perform the request won't have performance impact on the fluidity of the
interface.

The result of the request, the authentification token, was also stored
in the web browser session. This was changed in favor of having the
value stored in a state at the root of the brokerDetails component.

The only thing that is having a behavioral change since the previous
version is the case when the authentification is failing. Before the
user had an option to retrigger the authentification if that were the
case. Now the button are staying disable if the auth has failed.

We can easily work a solution for retriggering the login if needed.