Jolokia: store the token in the session
Use the web browser session to store the jolokia token.
The useJolokiaLogin hook connects to the api server only if there's no
token in the session cache. If there is one, it returns this value.

Lifecycle of the token:
If a request is made by the JolokiTestPanel to the api server and it
comes back with an error, the current token gets removed from the
session and the user is prompter to reload the page.

Procedure to  trigger such a situation:
* start the api server
* login to it
* restart the api server
* execute a query with the JolokiaTestPanel

You will get prompted to reload the page, upon reloading you'll see the
alert popup saying that you got successfully connected to Jolokia.