Implement PodsList Page with User-Selectable Broker Pod for Jolokia Endpoint Access
Before, the frontend was limited to only connect to the jolokia endpoint of the first pod.
It would have meant limitations to what information could be displayed in the future.

As a result, this commit changes the steps a user needs to take in order to display the broker details.
We now request that the user selects a pod prior to accessing the page.

This change paves the way for displaying pod-related information such as number of messages in flight in the queues.