[#128] brokerDetails: convert to react-query
The BrokerDetails page is now using the react-query's generated enpoint
to make its requests to the backend.

The hook useJolokiaLogin is checking on a regular interval the validity
of the token. When the token gets invalid, the hook will retrigger a
login to obtain a new token.

This is especially useful when the api-server gets restarted. This is
because the api-server keeps a mapping of a token to a Jolokia object.
Meaning that after a reboot a token could be valid but leading to a null
Jolokia object, and therefore to a crash.