ci: check that the generated files are in sync
Add a check in the CI to verify that the generated files are in sync
with the openapi.yml they depend on. This will fire an error if someone
either: forgets to update the files after a change in the openapi
definition or attempts to update the generated files manually.