[#146] api-server: fully qualified return types
Bring in fully qualified types for the swagger definition. Before
Generic types were used.

These had for main drawback to prevent us to generate proper types with
the codegen to handle data returned by the api-server. The data was
there but not named. So code like
`OperationResult.Op.listSessions[0].args` wasn't possible. We would have
needed to rely on json parsing. In turns meaning custom parsing code
here and there with a higher bug exposure. Now that we have fully
qualified names, we can rely on the generated typescript types to know
what is inside the data we get back from the server.

The api-server now needs to perform a few transformations to the data
from jolokia in order to be compliant with the return spec. In an ideal
world, the spec should have just needed to follow the jolokia output.
But some roadblock forced us to consider transforming the data in some
ways:
* Strings like `org.apache.activemq.artemis:broker="amq-broker` are
  transformed to a json counterpart:
  ```
  broker: {
      name: `amq-broker`
  }
  ```
* Operations, can be either an object or an array of objects in the
  jolokia response. The api-server transforms these into array of
  objects. This is necessary because there's a bug in the way the
  express-openapi-validator handles the `oneOf` property in the openapi
  definition.
To follow this change of behavior from the api-server, the tests needed
to be updated.

Last but not least, the api-server validates the output json against the
openapi definition too. So we're guaranteed that the result matches what
we expect it to be.