[#147] use a reducer to update the state of the CR
Updating the CR is a very complex task. It involves setting fields deep
down in certain places and some fields are expecting to trigger
cascading effect throughout the entire CR (none so far).

A reducer is a function that takes two parameters, a state to update and
a action to perform on to it.

An action takes an operation and an optional payload to execute.

All the valid operations are listed under ArtemisReducerActions. Follow
the switch case in the reducer to know which payload to associate with
which operation.

Operation can be simple setters like `setConsoleExpose` or named functions
like `incrementReplicas`.

The reducer is invoked with the `dispatch` function that can be
retrieved from the context:

```
const { yamlData } = useContext(BrokerConfigContext);
const dispatch = useContext(BrokerDispatchContext);
```

The `dispatch` is then invoked as follows:

```
dispatch({
  operation: ArtemisReducerActions.incrementReplicas,
})
```

All the reducer code is located in the `add-broker.ts` file that serves
as a single source of truth regarding how the CR state can be modified.

As a result the code of the components only focuses on displaying the UI
elements and calling the dispatcher to update the state. There is little
need for `useState` in the component themselves now, since the state is
handled directly using the reducer.