add-broker: variable naming overhaul
Some names used throughout the broker creation and update form where
lacking a bit of accuracy. This commit updates these:

- yamlData and brokerModel to formState

yamlData was too generic, brokerModel wasn't accurate since the data
structure was representing both data from the CR and information about
the state of the form like what is the editor.

- formState.yamlData to formState.cr

the yamlData field in the formState isn't really a yaml data, it's a JS
object that might get transformed to yaml at some point. It is in
contrast a CR that is being built.

- BrokerConfigContext to BrokerCreationFormState

The type of the object already convey the information that it is a
context. BrokerConfig was a bit too generic.

- BrokerDispatchContext to BrokerCreationFormDispatch

The type of the object already convey the information that it is a
context. BrokerDispatch was a bit too generic.