[#147] unit test the broker creation reducer
The broker creation reducer is a complex piece of code which is designed
to producing reproducible and accurate CRs. Therefore unit testing it is
paramount to ensure a predictable user experience. This commit adds a
testing file for that purpose, not all the tests have been written so
far  but the expectation is that it will be filled up in more details in
future additions.