[NOISSUE] use patternFly Form component
We can leverage the patternFly Form component at the root of the
acceptor, connector and console configuration page. The Form component
handles the layout for us and removes the need of flex organizers. Also
it's easier to have fields aligning properly without having to deal with
sideeffects when a subcomponent tries to get more space than it was
doing before.