[#150] presets instead of annotations
Use the wording `apply presets` instead of speaking of annotations to
the user. This is putting the focus on the functionality rather than on
the implementation details that the preset is applied via annotation.
The 'apply preset' link gets moved near the acceptor name, thus fixing
the confusion of having it in the middle of the props.