[NOISSUE] keep sslEnabled when changing secrets
When updating the secret, if the user had a -ptls secret (meaning they
where on a cert-manager preset most probably) the preset gets removed,
and thus goes away the sslEnabled flag. But in that particular case the
user wants to keep the sslEnabled on because they wanted to have the
secret name changed.