[#177] custom ui element for other parameters
Other parameters are a key value pair. The previous method was relying
on a textfield for the user to enter these value as a comma separated
list of properties. Now there's a ui component for it showing keys and
values in separated text fields. The user can add a new entry, edit it
at will and then update or delete it.

fixes #177