[NOISSUE] broker pods list, only show pods
There was a small issue with the broker pods list, in some instances it
could return a list of brokers instead of a list of pods. By removing
the call that was getting the list of brokers, this is fixed. I also
added empty states for when there's no pods in the broker for whatever
reasons.