[#193] update the sdk to version 1.3
There is a problem with our current setup, it was made to function on a
very outdated version of the dynamic plugin sdk (0.0.21). Running on dev
mode this wasn't an issue since the entire console was under our control
and matching the version number.

But once we build to prod, this is a different matter, the dynamic
plugin version that is running is the one from the cluster and has to
match. Therefore some of the calls we are running don't work at all.

This is the first step to debug the situation. Following commits will
address the diverse bugs we are experiencing in the UI.

See #193