[#211] remove storybook
It's been more than a year that our storybook is down. The issue raised
by the team at that time on the dyanmic plugin sdk has gone stale
https://github.com/openshift/dynamic-plugin-sdk/issues/216. Which means
that we can't hope for a resolution unless we actively engage with them
again on the subject.

Given the fact that there are other ways to test the code
(react-testing-library) and the fact that none of the current members of
the team know how to operate/leverage storybook. Let's just remove it.

This removes a lot of dependencies to external libraries and makes us
resemble more of a vanilla plugin, which in turn means easier
maintenance down the line.

fixes #211