[#240] move api server to its own repo
This commit removes the code from the api-server. The new repository
where the code lives now is
https://github.com/artemiscloud/activemq-artemis-jolokia-api-server.

The Dockerfile got synchronized back to the one from the
console-plugin-template at commit 2b4dbe4b4313e258be1754815614cf7d47f45c11

The deploy/ folder content has been synchronized back to its state
before introducing the api-server in the project. At commit
fc70c0ee0ad491a2bb5efb9c23d9cf1691cebde8.

The only difference left is that the plugin relies on the
activemq-artemis-jolokia-api-server proxy service to be there:

```
  proxy:
    - type: Service
      alias: api-server-service
      endpoint:
        type: Service
        service:
          name: activemq-artemis-jolokia-api-server
          namespace: activemq-artemis-jolokia-api-server
          port: 9443
```

Follow the updates in the Readme for more information about how to
develop and deploy now.

This work is linked to the PR that adds the jolokia-api-server code to
its own repo there
https://github.com/artemiscloud/activemq-artemis-jolokia-api-server/pull/4