[#198] the yaml editor allows to enter new Lines
Previously the yaml editor wasn't allowing a user to add new lines. This
was because at each keystroke the yaml file got formatted automatically,
removing what formatting felt as unnecessary lines.. This meant that the
user had to use tricks like copy&pasting to get to a state the wanted to
have.

To fix that, this commits reverts back to using the resource editor
provided by the console. It has several advantages:
* The use can edit the file as they want without bugs
* There's documentation included from the CRDs when hovering properties

The main drawback is that the user needs to click on save to save their
changes to the model, then needs to click on apply to apply their
config (either creating or updating)

Other additions:
----------------

In order to make the logic of the buttons the same between the two
views, they have been extracted up a notch in the component hierarchy.

Warning modals have been added to help the user know if they are about
to lose some changes.

A reload button was added to the update broker page, so that the user
can reload the content of the existing broker.