[#277] pf5: mitigating side effects
During the migration the appenTo functionality got broken somehow, this
could be due to the fact that there are some CSS missing on the console
atm, the bug has been notified to the openshift console team and is
actively worked on.

In the meantime, let's append to the entire document, which makes the
select work again.